// src/types/BankDatatTypes.ts
export interface BankCardsProps {
    banksData: BankData;
}
  
  export interface BankData {
    id: number;
    serviceCategoryName: string;
    serviceCategoryLogo: string | null;
    vdbServices: BankService[];
  }

  
  export interface BankService {
    id: number;
    serviceName: string;
    openType: 'new_tab' | 'iframe';
    serviceUrl: string;
    serviceImage: string | null;
    serviceVisibility:'0'| '1';
    serviceOpenType:'SIMPLE_MICROSITE' | 'SSO_LOGIN' | 'API_INTEGRATION';
  }
  